# Resolved Conflicts:
#	Modules/BookingManager/classes/class.ilObjBookingPool.php
#	Modules/BookingManager/classes/class.ilObjBookingPoolGUI.php
#	Modules/Category/classes/class.ilObjCategoryGUI.php
#	Modules/Course/classes/Export/class.ilCourseDefinedFieldDefinition.php
#	Modules/Course/classes/class.ilCourseRegistrationGUI.php
#	Modules/Course/classes/class.ilObjCourse.php
#	Modules/Course/classes/class.ilObjCourseAccess.php
#	Modules/Course/classes/class.ilObjCourseGUI.php
#	Modules/Course/classes/class.ilObjCourseGroupingGUI.php
#	Modules/Course/classes/class.ilObjCourseListGUI.php
#	Modules/DataCollection/classes/Content/class.ilDclContentImporter.php
#	Modules/Exercise/classes/class.ilExAssignment.php
#	Modules/Exercise/classes/class.ilExAssignmentEditorGUI.php
#	Modules/Exercise/classes/class.ilExSubmission.php
#	Modules/Exercise/classes/class.ilExerciseDataSet.php
#	Modules/Exercise/classes/class.ilObjExercise.php
#	Modules/Forum/classes/class.ilForumMailNotification.php
#	Modules/Group/classes/class.ilGroupParticipantsTableGUI.php
#	Modules/Group/classes/class.ilGroupRegistrationGUI.php
#	Modules/Group/classes/class.ilObjGroup.php
#	Modules/Group/classes/class.ilObjGroupAccess.php
#	Modules/Group/classes/class.ilObjGroupGUI.php
#	Modules/HTMLLearningModule/LuceneObjectDefinition.xml
#	Modules/Session/classes/class.ilObjSession.php
#	Modules/Session/classes/class.ilObjSessionAccess.php
#	Modules/Session/classes/class.ilObjSessionGUI.php
#	Modules/Survey/classes/class.ilObjSurvey.php
#	Modules/Survey/classes/class.ilObjSurveyGUI.php
#	Modules/Test/classes/class.ilObjTest.php
#	Modules/Test/classes/class.ilObjTestGUI.php
#	Modules/Test/classes/class.ilScoringAdjustmentGUI.php
#	Modules/Test/classes/class.ilTestEvaluationGUI.php
#	Modules/Test/classes/class.ilTestPlayerAbstractGUI.php
#	Modules/Test/classes/class.ilTestRandomQuestionSetConfigStateMessageHandler.php
#	Modules/Test/classes/class.ilTestScoringByQuestionsGUI.php
#	Modules/Test/classes/class.ilTestSubmissionReviewGUI.php
#	Modules/Test/classes/tables/class.ilTestFixedParticipantsTableGUI.php
#	Modules/Test/classes/tables/class.ilTestParticipantsTableGUI.php
#	Modules/TestQuestionPool/classes/class.assMatchingQuestionGUI.php
#	Modules/TestQuestionPool/classes/class.assOrderingQuestionGUI.php
#	Modules/TestQuestionPool/classes/class.assQuestionGUI.php
#	Modules/TestQuestionPool/classes/class.ilObjQuestionPool.php
#	Services/AccessControl/classes/class.ilRbacReview.php
#	Services/Authentication/classes/Provider/class.ilAuthProviderFactory.php
#	Services/Authentication/classes/class.ilAuthBase.php
#	Services/Authentication/classes/class.ilAuthUtils.php
#	Services/Authentication/classes/class.ilSessionStatistics.php
#	Services/COPage/css/content.css
#	Services/Component/classes/class.ilComponent.php
#	Services/Container/classes/class.ilContainer.php
#	Services/Container/classes/class.ilContainerGUI.php
#	Services/CopyWizard/classes/class.ilCopyWizardOptions.php
#	Services/Database/classes/class.ilDBUpdate.php
#	Services/Form/classes/class.ilTextWizardInputGUI.php
#	Services/GlobalCache/classes/class.ilGlobalCache.php
#	Services/InfoScreen/classes/class.ilInfoScreenGUI.php
#	Services/Init/classes/class.ilErrorHandling.php
#	Services/Init/classes/class.ilInitialisation.php
#	Services/Init/classes/class.ilStartUpGUI.php
#	Services/Init/templates/default/tpl.view_terms_of_service.html
#	Services/Language/classes/class.ilLanguage.php
#	Services/Locator/classes/class.ilLocatorGUI.php
#	Services/MainMenu/classes/class.ilMainMenuGUI.php
#	Services/MainMenu/templates/default/tpl.main_menu_list_entries.html
#	Services/Membership/classes/class.ilAttendanceList.php
#	Services/Membership/classes/class.ilMemberAgreementGUI.php
#	Services/Membership/classes/class.ilMembershipGUI.php
#	Services/Membership/classes/class.ilRegistrationGUI.php
#	Services/Membership/classes/class.ilWaitingListTableGUI.php
#	Services/Object/classes/class.ilObject.php
#	Services/Object/classes/class.ilObjectCopyGUI.php
#	Services/Object/classes/class.ilObjectGUI.php
#	Services/PrivacySecurity/classes/class.ilPrivacySettings.php
#	Services/Registration/classes/class.ilAccountRegistrationGUI.php
#	Services/Registration/classes/class.ilRegistrationCode.php
#	Services/Search/classes/class.ilSearchAutoComplete.php
#	Services/TermsOfService/classes/class.ilTermsOfServiceFileSystemDocument.php
#	Services/Tracking/classes/class.ilChangeEvent.php
#	Services/Tracking/classes/class.ilLPObjSettings.php
#	Services/Tracking/classes/class.ilLPStatus.php
#	Services/Tracking/classes/class.ilLearningProgressBaseGUI.php
#	Services/Tracking/classes/class.ilTrQuery.php
#	Services/Tree/classes/class.ilMaterializedPathTree.php
#	Services/User/classes/class.ilObjUserGUI.php
#	Services/User/classes/class.ilUserQuery.php
#	cron/classes/class.ilCronCheck.php
#	include/inc.ilias_version.php
#	lang/ilias_de.lang
#	setup/include/inc.setup_header.php
#	setup/sql/5_3_hotfixes.php
#	templates/default/delos.css
#	templates/default/delos.less
#	templates/default/less/variables.less

+++ Services/Utilities/classes/class.ilBenchmark.php
@@ -34,6 +34,21 @@ class ilBenchmark
 	protected $db_enabled_user;
 
 
+// fau: extendBenchmark - variables for instant saving
+    /**
+	 * Instant saving of benchmark to DB should be enabled
+     * @var boolean
+     */
+	protected $db_enabled_instant = false;
+
+
+    /**
+	 * Benchmark is currently saved to DB
+     * @var boolean
+     */
+	protected $db_saving_benchmark = false;
+// fau.
+
 	/**
 	 * Constructor
 	 */
@@ -107,18 +122,33 @@ class ilBenchmark
 		{
 			if (is_array($this->db_bench) && is_object($ilDB))
 			{
+// fau: extendBenchmark - store also the total time to compare it with query time
+				$settings = $DIC->settings();
+				$t1 = explode(" ", $GLOBALS['ilGlobalStartTime']);
+				$t2 = explode(" ", microtime());
+				$diff = $t2[0] - $t1[0] + $t2[1] - $t1[1];
+
+				$settings->set("db_bench_total_time", $diff);
+// fau.
+
 				$this->db_bench_stop_rec = true;
 
-				$ilDB->manipulate("DELETE FROM benchmark");
-				foreach ($this->db_bench as $b)
+// fau: extendBenchmark - check instant setting, add backtrace
+				if (empty($this->db_enabled_instant))
 				{
-					$id = $ilDB->nextId('benchmark');
-					$ilDB->insert("benchmark", array(
-						"id" => array("integer", $id),
-						"duration" => array("float", $this->microtimeDiff($b["start"], $b["stop"])),
-						"sql_stmt" => array("clob", $b["sql"])
-					));
+					$ilDB->manipulate("DELETE FROM benchmark");
+					foreach ($this->db_bench as $b)
+					{
+						$id = $ilDB->nextId('benchmark');
+						$ilDB->insert("benchmark", array(
+							"id" => array("integer", $id),
+							"duration" => array("float", $this->microtimeDiff($b["start"], $b["stop"])),
+							"sql_stmt" => array("text", $b["sql"]),
+							"backtrace" => array("text", $b["backtrace"])
+						));
+					}
 				}
+// fau.
 			}
 			$this->enableDbBench(false);
 		}
@@ -322,16 +352,20 @@ class ilBenchmark
 
 		$this->db_enabled = $ilSetting->get("enable_db_bench");
 		$this->db_enabled_user = $ilSetting->get("db_bench_user");
+// fau: extendBenchmark - get setting for instant saving
+		$this->db_enabled_instant =  $ilSetting->get("enable_db_bench_instant");
+// fau.
 		return $this->db_enabled;
 	}
 
+// fau: extendBenchmark - add parameter to enable instant saving, delete old data
 	/**
 	 * Enable DB benchmarking
 	 *
 	 * @param	boolean		enable db benchmarking
 	 * @param	string		user account name that should be benchmarked
 	 */
-	function enableDbBench($a_enable, $a_user = 0)
+	function enableDbBench($a_enable, $a_user = 0, $instant = 0)
 	{
 		global $DIC;
 		$ilSetting = $DIC->settings();
@@ -344,6 +378,8 @@ class ilBenchmark
 			{
 				$ilSetting->set("db_bench_user", $a_user);
 			}
+			$ilSetting->set("enable_db_bench_instant", $instant);
+			$this->clearData();
 		}
 		else
 		{
@@ -352,9 +388,10 @@ class ilBenchmark
 			{
 				$ilSetting->set("db_bench_user", $a_user);
 			}
+			$ilSetting->set("enable_db_bench_instant", $instant);
 		}
 	}
-
+// fau.
 
 	/**
 	 * start measurement
@@ -398,13 +435,42 @@ class ilBenchmark
 		if ($this->isDbBenchEnabled() && is_object($ilUser)
 		    && $this->db_enabled_user == $ilUser->getLogin()
 		    && !$this->db_bench_stop_rec) {
-			$this->db_bench[] = array(
-				"start" => $this->start,
-				"stop"  => microtime(),
-				"sql"   => $this->sql,
-			);
+// fau: extendBenchmark - add backtrace to benchmark, save instantly (to allow benchmarking of fatal errors)
+			$i = 0;
+			foreach (debug_backtrace() as $step)
+			{
+	            if ($i > 0)
+				{
+					$backtrace .= '['.$i.'] '.$step['file'].' '.$step['line'].': '.$step['function']."()\n";
+				}
+				$i++;
+			}
+			$backtrace .= '['.$i.'] '.$_SERVER['REQUEST_URI'];
 
-			return true;
+			if ($this->db_enabled_instant)
+			{
+				// prevent infinite recursion
+				$this->db_bench_stop_rec = true;
+
+				$ilDB = $DIC->database();
+				$id = $ilDB->nextId('benchmark');
+				$ilDB->insert("benchmark", array(
+					"id" => array("integer", $id),
+					"duration" => array("float", $this->microtimeDiff($this->start,  microtime())),
+					"sql_stmt" => array("text", $this->sql),
+					"backtrace" => array("text", $backtrace)
+				));
+
+				$this->db_bench_stop_rec  = false;
+			}
+
+			$this->db_bench[] = array(
+					"start" => $this->start,
+					"stop" => microtime(),
+					"sql" => $this->sql,
+					"backtrace" => $backtrace
+			);
+// fau.
 		}
 
 		return false;
@@ -425,12 +491,35 @@ class ilBenchmark
 		$b = array();
 		while ($rec = $ilDB->fetchAssoc($set))
 		{
+// fau: extendBenchmark - get benchmarks with backtrace
 			$b[] = array("sql" => $rec["sql_stmt"],
-				"time" => $rec["duration"]);
+				"time" => $rec["duration"],
+				"backtrace" => $rec["backtrace"]);
+// fau.
 		}
 		return $b;
 	}
 
+// fau: extendBenchmark - get total and query time
+
+	function getDbBenchTotalTime()
+	{
+		global $ilias;
+		return $ilias->getSetting('db_bench_total_time');
+	}
+
+	function getDbBenchQueryTime()
+	{
+		global $ilDB;
+
+		$set = $ilDB->query("SELECT SUM(duration) dsum FROM benchmark");
+		if ($rec = $ilDB->fetchAssoc($set))
+		{
+			return $rec['dsum'];
+		}
+		return 0;
+	}
+// fau.
 }
 
 ?>
